% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_amenities.R
\name{get_amenities}
\alias{get_amenities}
\title{Retrieve amenities for a neighbourhood/location}
\usage{
get_amenities(location, amenities = NULL)
}
\arguments{
\item{location}{The bbox of the location as determined by \code{nearby()}.}

\item{amenities}{a character vector representing amentities to add as
defined by \href{https://wiki.openstreetmap.org/wiki/Key:amenity}{Open Street Map}.
If NULL will default to all amenity tags. For valid amenity tags use
\code{osmdata::available_tags("amenity")}}
}
\value{
an sf Simple Features Collection of points
}
\description{
Retrieve amenities for a neighbourhood/location
}
