% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roads.R
\name{get_roads}
\alias{get_roads}
\title{Get the road network for a location}
\usage{
get_roads(location, ...)
}
\arguments{
\item{location}{A character string specifying the location of interest (e.g. "Melbourne Vic") or a matrix-like bounding box.}

\item{...}{Extra arguments to the function \code{opq}.}
}
\value{
An object of class sf
}
\description{
This function takes a location specified by the user and returns a \code{sf} object which contains all walkable roads in and around the location.
}
\details{
A walkable road is defined as a road which (1) does not \emph{not have sidewalks} and (2) is \emph{not listed as a motorway or a motorway link} in OpenStreetMaps.
}
\examples{

\dontrun{
# For Melbourne, Victoria, Australia

melbourne <- get_roads("melbourne vic")
}
}
