% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roads_isochrone.R
\name{get_roads_isochrone}
\alias{get_roads_isochrone}
\title{Get buffered road network}
\usage{
get_roads_isochrone(nearby_obj, buffer_dist = 2e-04)
}
\arguments{
\item{nearby_obj}{An object of class \code{nearby}.}

\item{buffer_dist}{A buffer distance in decimal degrees. Defaults to 0.0002 (~20 metres).}
}
\value{
An object of class sf containing polygonal features representing the parts of the road network that lead to accessible amenities.
}
\description{
Get polygons representing the roads leading to the accessible amenities returned by \code{nearby()}
}
