% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby.R
\name{nearby}
\alias{nearby}
\title{Is there stuff nearby?}
\usage{
nearby(location, radius = 1500, amenities = NULL)
}
\arguments{
\item{location}{where to compute walkability}

\item{radius}{walkable distance in meters}

\item{amenities}{character vector of amenities to include
(or with minus sign exclude) from calculation. The default
is to include all of them}
}
\description{
Given a location, look up features within a walking distance of 'radius' to summarise walkability
}
