% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-method.R
\name{plot.nearby}
\alias{plot.nearby}
\title{Produce a map of nearby amenities}
\usage{
\method{plot}{nearby}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{nearby}}

\item{y}{Not used for \code{plot.nearby}}

\item{...}{Additional arguments to plot, currently does not do anything}
}
\value{
a Leaflet map widget
}
\description{
Produce a map of nearby amenities
}
\details{
The plot method produces a map of nearby amenities, with
a convex hull over the amenities' longitude and latitude. There is
a dropdown menu on the top right, to filter by by an amenity type.
}
\examples{
\dontrun{
plot(nearby("Carlton North, Victoria")) 
plot(nearby("Carlton North, Victoria", radius = walk_time(15))) 
}

}
