% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{walk_time}
\alias{walk_time}
\title{Convert time to distance}
\usage{
walk_time(minutes, speed=c("medium", "slow","fast"))
}
\arguments{
\item{minutes}{time spent walking}

\item{speed}{how fast you walk, either "slow", "medium", "fast" or a numeric value in metres per second}
}
\description{
The speeds are 0.8 m/s for 'slow', 1.4 m/s for 'medium' and 2.5m/s for 'fast'
}
