% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_amenities.R
\name{get_amenities}
\alias{get_amenities}
\title{Retrieve amenities for a neighbourhood}
\usage{
get_amenities(location, ..., amenities = NULL)
}
\arguments{
\item{location}{A character string or bounding box as passed to \code{\link[osmdata:opq]{osmdata::opq()}}.}

\item{...}{additional arguments to pass to \code{\link[osmdata:opq]{osmdata::opq()}}.}

\item{amenities}{a character vector representing amentities to add as
defined by \href{https://wiki.openstreetmap.org/wiki/Key:amenity}{Open Street Map}.
If NULL will default to all amenity tags.}
}
\value{
an sf Simple Features Collection of points
}
\description{
Retrieve amenities for a neighbourhood
}
