% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roads_graph.R
\name{get_roads_graph}
\alias{get_roads_graph}
\title{Get the sf object for a road network}
\usage{
get_roads_graph(location)
}
\arguments{
\item{location}{A character string specifying the location of interest (e.g. "Melbourne Vic") or a matrix-like bounding box.}
}
\value{
An object of class data.frame/dodgr_streetnet
}
\description{
This function takes a location and returns an \code{sf} object which contains the walkable roads in and around that location.
}
\examples{

\dontrun{
# For Moorooka, Brisbane, QLD

get_roads_graph("moorooka, brisbane, qld")
}
}
